daemon,o/devicestate: move user creation and removal helpers to o/devicestate
