+++ daemon/api_users.go
@@ -21,25 +21,15 @@ package daemon
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
-	"os/user"
-	"path/filepath"
 	"regexp"
-	"time"
 
-	"github.com/snapcore/snapd/asserts"
 	"github.com/snapcore/snapd/client"
-	"github.com/snapcore/snapd/logger"
-	"github.com/snapcore/snapd/osutil"
-	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/configstate/config"
-	"github.com/snapcore/snapd/overlord/snapstate"
-	"github.com/snapcore/snapd/overlord/state"
+	"github.com/snapcore/snapd/overlord/devicestate"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/store"
-	"github.com/snapcore/snapd/strutil"
 )
 
 var (
@@ -72,8 +62,8 @@ var (
 )
 
 var (
-	osutilAddUser = osutil.AddUser
-	osutilDelUser = osutil.DelUser
+	deviceStateCreateUser = devicestate.CreateUser
+	deviceStateRemoveUser = devicestate.RemoveUser
 )
 
 // userResponseData contains the data releated to user creation/login/query
@@ -233,36 +223,13 @@ func postUsers(c *Command, r *http.Request, user *auth.UserState) Response {
 }
 
 func removeUser(c *Command, username string, opts postUserDeleteData) Response {
-	// TODO: allow to remove user entries by email as well
-
-	// catch silly errors
-	if username == "" {
-		return BadRequest("need a username to remove")
-	}
-	// check the user is known to snapd
-	st := c.d.overlord.State()
-	st.Lock()
-	_, err := auth.UserByUsername(st, username)
-	st.Unlock()
-	if err == auth.ErrInvalidUser {
-		return BadRequest("user %q is not known", username)
-	}
+	u, internal, err := deviceStateRemoveUser(c.d.overlord.State(), username)
 	if err != nil {
-		return InternalError(err.Error())
-	}
-
-	// first remove the system user
-	if err := osutilDelUser(username, &osutil.DelUserOptions{ExtraUsers: !release.OnClassic}); err != nil {
-		return InternalError(err.Error())
-	}
-
-	// then the UserState
-	st.Lock()
-	u, err := auth.RemoveUserByUsername(st, username)
-	st.Unlock()
-	// ErrInvalidUser means "not found" in this case
-	if err != nil && err != auth.ErrInvalidUser {
-		return InternalError(err.Error())
+		if internal {
+			return InternalError(err.Error())
+		} else {
+			return BadRequest(err.Error())
+		}
 	}
 
 	result := map[string]interface{}{
@@ -292,6 +259,7 @@ func postCreateUser(c *Command, r *http.Request, user *auth.UserState) Response 
 }
 
 func createUser(c *Command, createData postUserCreateData) Response {
+	var createdUsersResponse []userResponseData
 	// verify request
 	st := c.d.overlord.State()
 	st.Lock()
@@ -335,196 +303,23 @@ func createUser(c *Command, createData postUserCreateData) Response {
 		createData.Sudoer = true
 	}
 
-	var model *asserts.Model
-	var serial *asserts.Serial
-	createKnown := createData.Known
-	if createKnown {
-		var err error
-		st.Lock()
-		model, err = c.d.overlord.DeviceManager().Model()
-		st.Unlock()
-		if err != nil {
-			return InternalError("cannot create user: cannot get model assertion: %v", err)
-		}
-		st.Lock()
-		serial, err = c.d.overlord.DeviceManager().Serial()
-		st.Unlock()
-		if err != nil && err != state.ErrNoState {
-			return InternalError("cannot create user: cannot get serial: %v", err)
-		}
-	}
-
-	// special case: the user requested the creation of all known
-	// system-users
-	if createData.Email == "" && createKnown {
-		return createAllKnownSystemUsers(st, model, serial, &createData)
-	}
-	if createData.Email == "" {
-		return BadRequest("cannot create user: 'email' field is empty")
-	}
-
-	var username string
-	var opts *osutil.AddUserOptions
-	if createKnown {
-		username, opts, err = getUserDetailsFromAssertion(st, model, serial, createData.Email)
-	} else {
-		username, opts, err = getUserDetailsFromStore(storeFrom(c.d), createData.Email)
-	}
+	createdUsers, internal, err := deviceStateCreateUser(st, c.d.overlord.DeviceManager(), createData.Sudoer, createData.Known, createData.Email)
 	if err != nil {
-		return BadRequest("%s", err)
-	}
-
-	// FIXME: duplicated code
-	opts.Sudoer = createData.Sudoer
-	opts.ExtraUsers = !release.OnClassic
-
-	if err := osutilAddUser(username, opts); err != nil {
-		return BadRequest("cannot create user %s: %s", username, err)
-	}
-
-	if err := setupLocalUser(c.d.overlord.State(), username, createData.Email); err != nil {
-		return InternalError("%s", err)
-	}
-
-	result := userResponseData{
-		Username: username,
-		SSHKeys:  opts.SSHKeys,
-	}
-
-	if createData.singleUserResultCompat {
-		// return a single userResponseData in this case
-		return SyncResponse(&result)
-	} else {
-		return SyncResponse([]userResponseData{result})
-	}
-}
-
-func getUserDetailsFromStore(theStore snapstate.StoreService, email string) (string, *osutil.AddUserOptions, error) {
-	v, err := theStore.UserInfo(email)
-	if err != nil {
-		return "", nil, fmt.Errorf("cannot create user %q: %s", email, err)
-	}
-	if len(v.SSHKeys) == 0 {
-		return "", nil, fmt.Errorf("cannot create user for %q: no ssh keys found", email)
-	}
-
-	// Amend information where the key came from to ensure it can
-	// be update/replaced later
-	for i, k := range v.SSHKeys {
-		v.SSHKeys[i] = fmt.Sprintf(`%s # snapd {"origin":"store","email":%q}`, k, email)
-	}
-
-	gecos := fmt.Sprintf("%s,%s", email, v.OpenIDIdentifier)
-	opts := &osutil.AddUserOptions{
-		SSHKeys: v.SSHKeys,
-		Gecos:   gecos,
-	}
-	return v.Username, opts, nil
-}
-
-func createAllKnownSystemUsers(st *state.State, modelAs *asserts.Model, serialAs *asserts.Serial, createData *postUserCreateData) Response {
-	var createdUsers []userResponseData
-	headers := map[string]string{
-		"brand-id": modelAs.BrandID(),
-	}
-
-	st.Lock()
-	db := assertstate.DB(st)
-	assertions, err := db.FindMany(asserts.SystemUserType, headers)
-	st.Unlock()
-	if err != nil && !asserts.IsNotFound(err) {
-		return BadRequest("cannot find system-user assertion: %s", err)
-	}
-
-	for _, as := range assertions {
-		email := as.(*asserts.SystemUser).Email()
-		// we need to use getUserDetailsFromAssertion as this verifies
-		// the assertion against the current brand/model/time
-		username, opts, err := getUserDetailsFromAssertion(st, modelAs, serialAs, email)
-		if err != nil {
-			logger.Noticef("ignoring system-user assertion for %q: %s", email, err)
-			continue
+		if internal {
+			return InternalError(err.Error())
+		} else {
+			return BadRequest(err.Error())
 		}
-		// ignore already existing users
-		if _, err := userLookup(username); err == nil {
-			continue
-		}
+	}
 
-		// FIXME: duplicated code
-		opts.Sudoer = createData.Sudoer
-		opts.ExtraUsers = !release.OnClassic
-
-		if err := osutilAddUser(username, opts); err != nil {
-			return InternalError("cannot add user %q: %s", username, err)
-		}
-		if err := setupLocalUser(st, username, email); err != nil {
-			return InternalError("%s", err)
-		}
-		createdUsers = append(createdUsers, userResponseData{
-			Username: username,
-			SSHKeys:  opts.SSHKeys,
+	for _, cu := range createdUsers {
+		createdUsersResponse = append(createdUsersResponse, userResponseData{
+			Username: cu.Username,
+			SSHKeys:  cu.SSHKeys,
 		})
 	}
 
-	return SyncResponse(createdUsers)
-}
-
-func getUserDetailsFromAssertion(st *state.State, modelAs *asserts.Model, serialAs *asserts.Serial, email string) (string, *osutil.AddUserOptions, error) {
-	errorPrefix := fmt.Sprintf("cannot add system-user %q: ", email)
-
-	st.Lock()
-	db := assertstate.DB(st)
-	st.Unlock()
-
-	brandID := modelAs.BrandID()
-	series := modelAs.Series()
-	model := modelAs.Model()
-
-	a, err := db.Find(asserts.SystemUserType, map[string]string{
-		"brand-id": brandID,
-		"email":    email,
-	})
-	if err != nil {
-		return "", nil, fmt.Errorf(errorPrefix+"%v", err)
-	}
-	// the asserts package guarantees that this cast will work
-	su := a.(*asserts.SystemUser)
-
-	// cross check that the assertion is valid for the given series/model
-
-	// check that the signer of the assertion is one of the accepted ones
-	sysUserAuths := modelAs.SystemUserAuthority()
-	if len(sysUserAuths) > 0 && !strutil.ListContains(sysUserAuths, su.AuthorityID()) {
-		return "", nil, fmt.Errorf(errorPrefix+"%q not in accepted authorities %q", email, su.AuthorityID(), sysUserAuths)
-	}
-	if len(su.Series()) > 0 && !strutil.ListContains(su.Series(), series) {
-		return "", nil, fmt.Errorf(errorPrefix+"%q not in series %q", email, series, su.Series())
-	}
-	if len(su.Models()) > 0 && !strutil.ListContains(su.Models(), model) {
-		return "", nil, fmt.Errorf(errorPrefix+"%q not in models %q", model, su.Models())
-	}
-	if len(su.Serials()) > 0 {
-		if serialAs == nil {
-			return "", nil, fmt.Errorf(errorPrefix + "bound to serial assertion but device not yet registered")
-		}
-		serial := serialAs.Serial()
-		if !strutil.ListContains(su.Serials(), serial) {
-			return "", nil, fmt.Errorf(errorPrefix+"%q not in serials %q", serial, su.Serials())
-		}
-	}
-
-	if !su.ValidAt(time.Now()) {
-		return "", nil, fmt.Errorf(errorPrefix + "assertion not valid anymore")
-	}
-
-	gecos := fmt.Sprintf("%s,%s", email, su.Name())
-	opts := &osutil.AddUserOptions{
-		SSHKeys:             su.SSHKeys(),
-		Gecos:               gecos,
-		Password:            su.Password(),
-		ForcePasswordChange: su.ForcePasswordChange(),
-	}
-	return su.Username(), opts, nil
+	return SyncResponse(createdUsersResponse)
 }
 
 type postUserData struct {
@@ -549,52 +344,6 @@ type postUserCreateData struct {
 }
 
 type postUserDeleteData struct{}
-
-var userLookup = user.Lookup
-
-func setupLocalUser(st *state.State, username, email string) error {
-	user, err := userLookup(username)
-	if err != nil {
-		return fmt.Errorf("cannot lookup user %q: %s", username, err)
-	}
-	uid, gid, err := osutil.UidGid(user)
-	if err != nil {
-		return err
-	}
-	authDataFn := filepath.Join(user.HomeDir, ".snap", "auth.json")
-	if err := osutil.MkdirAllChown(filepath.Dir(authDataFn), 0700, uid, gid); err != nil {
-		return err
-	}
-
-	// setup new user, local-only
-	st.Lock()
-	authUser, err := auth.NewUser(st, username, email, "", nil)
-	st.Unlock()
-	if err != nil {
-		return fmt.Errorf("cannot persist authentication details: %v", err)
-	}
-	// store macaroon auth, user's ID, email and username in auth.json in
-	// the new users home dir
-	outStr, err := json.Marshal(struct {
-		ID       int    `json:"id"`
-		Username string `json:"username"`
-		Email    string `json:"email"`
-		Macaroon string `json:"macaroon"`
-	}{
-		ID:       authUser.ID,
-		Username: authUser.Username,
-		Email:    authUser.Email,
-		Macaroon: authUser.Macaroon,
-	})
-	if err != nil {
-		return fmt.Errorf("cannot marshal auth data: %s", err)
-	}
-	if err := osutil.AtomicWriteFileChown(authDataFn, []byte(outStr), 0600, 0, uid, gid); err != nil {
-		return fmt.Errorf("cannot write auth file %q: %s", authDataFn, err)
-	}
-
-	return nil
-}
 
 func getUsers(c *Command, r *http.Request, user *auth.UserState) Response {
 	st := c.d.overlord.State()
